Split Qry into Qry and Expr
While `Qry` still provides a monadic interface, most of the actual calculation
happens in the interpreter for `Expr`, and any past-horizon exceptions come
from the evaluation of an `Expr`. Unlike `Qry`, `Expr` uses PHOAS, which means
that we can actually `Show` expressions, despite them containing functions.
This means that we can include the offending expression in the exception.

This is also a strict generalization: previously, an entire `Qry` was always
evaluated against a single era; with the split, every `Expr` is evaluated
against a single era, but every `Expr` embedded in a `Qry` can be evaluated
against a _different_ era.

Closes #2463.